before alternate download implementation

+++ schema/util/CALC_DIR_DEG.sql
@@ -1,5 +1,5 @@
 CREATE OR REPLACE FUNCTION calc_dir_deg(u FLOAT, v FLOAT)
-  RETURNS FLOAT AS 'SELECT (degrees(atan2(u, v))::int + 360) % 360;'
+  RETURNS INT AS 'SELECT (degrees(atan2(u, v))::int + 360) % 360;'
   LANGUAGE SQL
   IMMUTABLE
   RETURNS NULL ON NULL INPUT;
